backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
python-dateutil==2.8.2
pandas==2.0.3
requests==2.31.0 